Add files via upload
The image for the landing section/area was reduced in size, to fit the view-port height, blurred and a white colour layer was added on top. The <main> section was also positioned close to the bottom of the page. 

A parallax effect was also added to the images on both sections.

Added a second section to the page, with coffee cup icons, with the first two tooltips appearing on hover of the first two icons.